Update to TWSAPI 971.01. Tests pass against IB Gateway 943.2a.

+++ combo_leg.go
@@ -23,5 +23,5 @@ type ComboLeg struct {
 	OpenClose          int64
 	ShortSaleSlot      int64
 	DesignatedLocation string
-	ExemptCode         int
+	ExemptCode         int64
 }
